fix(core): some small fixes
refs. metriport/metriport-internal#2502
