+++ packages/core/src/external/fhir/normalization/normalize-fhir.ts
